requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
Django==3.1.1
django-filter==2.3.0
django-graphql-jwt==0.3.1
django-storages==1.10
djangorestframework==3.11.1
gcloud==0.18.3
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
oauth2client==4.1.3
Pillow==7.2.0
promise==2.3
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
rsa==4.6
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.10
